RF_02.003.01.01 Перенес тест test_access_scm_title из test_freestyle_project.py в test_freestyle_project_configuration.py Вместе с переменными и функцией wait_until_clickable
