+++ tests/test_freestyle_project_configuration.py
@@ -1,10 +1,20 @@
+from selenium.webdriver.remote.webdriver import WebDriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 
+FREESTYLE_PROJECT_NAME = "Freestyle Project"
+SCM_TITLE_EXPECTED = "Source Code Management"
+
+
+def wait_until_clickable(browser: WebDriver, locator: tuple[str, str], timeout=10):
+    """Ожидает кликабельности элемента и возвращает его."""
+    return WebDriverWait(browser, timeout).until(
+        EC.element_to_be_clickable(locator)
+    )
+
 
 def create_freestyle_project(driver):
-
     driver.find_element(By.XPATH, "//a[@href='/view/all/newJob']").click()
     driver.find_element(By.ID, "name").send_keys('Test')
     driver.find_element(By.CSS_SELECTOR, "li.hudson_model_FreeStyleProject").click()
@@ -32,3 +42,15 @@ def test_disable_active_project(browser):
     assert browser.find_element(
         By.XPATH, "//*[contains(text(), 'This project is currently disabled')]"
     )
+
+
+def test_access_scm_title(browser):
+    browser.find_element(By.XPATH, "//a[@href='newJob']").click()
+
+    browser.find_element(By.ID, "name").send_keys(FREESTYLE_PROJECT_NAME)
+    browser.find_element(By.CLASS_NAME, "hudson_model_FreeStyleProject").click()
+    wait_until_clickable(browser, (By.ID, "ok-button")).click()
+
+    scm_title_text = wait_until_clickable(browser, (By.ID, "source-code-management")).text
+
+    assert scm_title_text == SCM_TITLE_EXPECTED
